requirements.txt
# Enhanced Energy Management System for GoodWe Inverter
# Core dependencies

# GoodWe inverter communication library
goodwe==0.4.8

# Configuration and data handling
PyYAML>=6.0

# Data analysis and processing
pandas>=1.3.0
numpy>=1.21.0

# HTTP requests for API calls
requests>=2.25.0
aiohttp>=3.8.0

# Date and time handling
python-dateutil>=2.8.0

# Logging and monitoring
colorlog>=6.0.0

# Optional: For advanced data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web server for remote log access
flask>=2.0.0
flask-cors>=3.0.0
psutil>=5.8.0

# Database dependencies
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
backoff>=2.2.0

# Development and testing
pytest>=6.0.0
pytest-asyncio>=0.15.0
black>=21.0.0
flake8>=3.9.0
